24년 8월 13일 1차 수정본

+++ src/main/java/org/zerock/chain/domain/EmployeesEntity.java
@@ -17,7 +17,7 @@ public class EmployeesEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "emp_no")
-    private Integer empNo;
+    private Long empNo;
 
     @Column(name = "first_name")
     private String firstName;
